add script that can create single site inputs
This has always been a missing piece so to speak.
But now finally figured out how to appropriately
deal with the projection info (keep it in wgs84) and
this gets around the issue that gdal had wiht cropping
out super small areas.
Annoying because of how redundant this code is with the
create_region_input.py script. Oh well.